+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -23,7 +23,6 @@ const
 	isUp = is('up'),
 	JS = 'JS',
 	Native = 'Native',
-	isItemDisabledDefault = () => (false),
 	// using 'bitwise or' for string > number conversion based on performance: https://jsperf.com/convert-string-to-number-techniques/7
 	getNumberValue = (index) => index | 0;
 
@@ -114,32 +113,6 @@ const VirtualListBaseFactory = (type) => {
 			 */
 			initUiChildRef: PropTypes.func,
 
-			/**
-			 * The Function that returns `true` if the item at the index is disabled.
-			 * It is used to navigate a list properly with 5 way keys, page up key,
-			 * and page down key. If it is not supplied, it assumes that no items are disabled.
-			 *
-			 * Usage:
-			 * ```
-			 * isItemDisabled = (index) => (this.items[index].disabled)
-			 * render = () => {
-			 * 	return (
-			 * 		<VirtualList
-			 * 			dataSize={this.items.length}
-			 * 			isItemDisabled={isItemDisabled}
-			 * 			itemRenderer={this.renderItem}
-			 * 			itemSize={this.itemSize}
-			 * 		/>
-			 * 	);
-			 * }
-			 * ```
-			 *
-			 * @type {Function}
-			 * @param {Number} index
-			 * @public
-			 */
-			isItemDisabled: PropTypes.func,
-
 			/*
 			 * It scrolls by page when `true`, by item when `false`.
 			 *
@@ -195,7 +168,6 @@ const VirtualListBaseFactory = (type) => {
 		static defaultProps = {
 			animate: false,
 			dataSize: 0,
-			isItemDisabled: isItemDisabledDefault,
 			pageScroll: false,
 			spacing: 0,
 			wrap: false
@@ -346,26 +318,16 @@ const VirtualListBaseFactory = (type) => {
 
 		findSpottableItem = (indexFrom, indexTo) => {
 			const
-				{dataSize, isItemDisabled} = this.props,
+				{dataSize} = this.props,
 				safeIndexFrom = clamp(0, dataSize - 1, indexFrom),
 				safeIndexTo = clamp(-1, dataSize, indexTo),
 				delta = (indexFrom < indexTo) ? 1 : -1;
 
 			if (indexFrom < 0 && indexTo < 0 || indexFrom >= dataSize && indexTo >= dataSize) {
 				return -1;
-			} else if (isItemDisabled === isItemDisabledDefault) {
+			} else {
 				return safeIndexFrom;
 			}
-
-			if (safeIndexFrom !== safeIndexTo) {
-				for (let i = safeIndexFrom; i !== safeIndexTo; i += delta) {
-					if (!isItemDisabled(i)) {
-						return i;
-					}
-				}
-			}
-
-			return -1;
 		}
 
 		findSpottableItemWithPositionInExtent = (indexFrom, indexTo, position) => {
@@ -377,7 +339,6 @@ const VirtualListBaseFactory = (type) => {
 				-1 <= indexTo && indexTo <= dataSize &&
 				0 <= position && position < dimensionToExtent) {
 				const
-					{isItemDisabled} = this.props,
 					direction = (indexFrom < indexTo) ? 1 : -1,
 					delta = direction * dimensionToExtent,
 					diffPosition = (indexFrom % dimensionToExtent) - position,
@@ -385,11 +346,7 @@ const VirtualListBaseFactory = (type) => {
 					// When direction is -1 (backward) and diffPosition is negative, substract dimensionToExtent.
 					startIndex = indexFrom - diffPosition + ((direction * diffPosition > 0) ? delta : 0);
 
-				for (let i = startIndex; direction * (indexTo - i) > 0; i += delta) {
-					if (!isItemDisabled(i)) {
-						return i;
-					}
-				}
+				return startIndex;
 			}
 
 			return -1;
@@ -414,7 +371,7 @@ const VirtualListBaseFactory = (type) => {
 
 		findNearestSpottableItemInExtent = (index, extentIndex) => {
 			const
-				{dataSize, isItemDisabled} = this.props,
+				{dataSize} = this.props,
 				{dimensionToExtent} = this.uiRefCurrent,
 				currentPosInExtent = clamp(0, dataSize - 1, index) % dimensionToExtent,
 				firstIndexInExtent = clamp(0, this.getExtentIndex(dataSize - 1), extentIndex) * dimensionToExtent,
@@ -425,12 +382,10 @@ const VirtualListBaseFactory = (type) => {
 				nearestIndex = -1;
 
 			for (let i = firstIndexInExtent; i < lastIndexInExtent; ++i) {
-				if (!isItemDisabled(i)) {
-					distance = Math.abs(currentPosInExtent - i % dimensionToExtent);
-					if (distance < minDistance) {
-						minDistance = distance;
-						nearestIndex = i;
-					}
+				distance = Math.abs(currentPosInExtent - i % dimensionToExtent);
+				if (distance < minDistance) {
+					minDistance = distance;
+					nearestIndex = i;
 				}
 			}
 
@@ -837,7 +792,6 @@ const VirtualListBaseFactory = (type) => {
 
 			delete rest.animate;
 			delete rest.initUiChildRef;
-			delete rest.isItemDisabled;
 			delete rest.spotlightId;
 			delete rest.wrap;